AI-NLP/speech-to-text/requirements.txt
absl-py==0.14.1
asgiref==3.4.1
astor==0.8.1
atomicwrites==1.4.0
attrs==21.2.0
audioread==2.1.9
backports.entry-points-selectable==1.1.0
cached-property==1.5.2
cfgv==3.3.1
colorama==0.4.4
cycler==0.10.0
decorator==5.1.0
distlib==0.3.3
Django==3.2.5
django-cors-headers==3.8.0
djangorestframework==3.12.4
filelock==3.3.0
gast==0.2.2
google-pasta==0.2.0
grpcio==1.41.0
h5py==3.4.0
identify==2.3.0
importlib-metadata==4.8.1
iniconfig==1.1.1
joblib==0.13.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
librosa==0.6.3
llvmlite==0.31.0
Markdown==3.3.4
matplotlib==3.4.3
mpi4py==3.1.1
nltk==3.2.5
nodeenv==1.6.0
numba==0.48.0
numpy==1.21.2
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.3
Pillow==8.3.2
platformdirs==2.4.0
pluggy==1.0.0
portalocker==2.3.2
pre-commit==2.15.0
protobuf==3.18.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
python-speech-features==0.6
pytz==2021.1
PyYAML==5.4.1
regex==2021.10.8
resampy==0.2.2
sacrebleu==2.0.0
scikit-learn==1.0
scipy==1.7.1
sentencepiece==0.1.96
six==1.16.0
sqlparse==0.4.2
tabulate==0.8.9
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
threadpoolctl==3.0.0
toml==0.10.2
tqdm==4.62.3
typing-extensions==3.10.0.2
virtualenv==20.8.1
Werkzeug==2.0.2